% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{bounds}
\alias{bounds}
\title{Set element-wise weight bounds}
\usage{
bounds(model, min = 0, max = Inf)
}
\arguments{
\item{model}{A \code{raking_model}.}

\item{min}{Lower bound on weights (>= 0). Default 0.}

\item{max}{Upper bound on weights (> 0). Default Inf.}
}
\value{
The updated \code{raking_model}.
}
\description{
Set element-wise weight bounds
}
\examples{
raking() |> bounds(min = 0.5, max = 2)
}
