% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/byType.R
\name{byType}
\alias{byType}
\title{Return indicator matrix}
\usage{
byType(dat, types, reqs)
}
\arguments{
\item{dat}{a dataframe representing a single categorical or numerical group to be processed}

\item{types}{representing the types of the input variable, possible values: "factor", "numeric","logic". 
if factor: recode the factors into different groups
if numeric: cut the numeric variables into factors
if logic(indicator matrix): group different factors if needed}

\item{reqs}{rules as string for recode/cut/group based on the different types of input variable 
if factor: examples as "c(0)='No';c(1,2,3,4)='Yes'", recode 0 into 'No', recode 1,2,3,4 into 'Yes'
if numeric: input the cuts
if logic(indicator matrix): input the categories to be grouped together}
}
\value{
A dataframe of the indicator matrix
}
\description{
This function takes in a dataframe representing a single categorical/numerical group.
based on the different types of data the input dataframe representing, 
it will perform different processure like cut, recode and etc and return a indicator matirx
}
