% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInteraction.R
\name{getInteraction}
\alias{getInteraction}
\title{Helper function for shinyapp}
\usage{
getInteraction(dat, inters)
}
\description{
This function get the interactions of two variables in the logic form
}
