% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSim.R
\name{getSim}
\alias{getSim}
\title{Return resulting percents/variance/mean/CIs for education level and citybelonging levels}
\usage{
getSim(surveyInput, censusInput, sample, seed, biased)
}
\arguments{
\item{surveyInput}{dataframe of individuals survey data containing gender, race, age info}

\item{censusInput}{dataframe of census constrain data, each column represents a census category}

\item{sample}{number of samples to take}

\item{seed}{seeds to be used for sampling}

\item{biased}{if biased > 0, use biased as the prob of sampling female individuals}
}
\value{
A list of the resulting percentages in the samples, the mean/variance/CIs for education levels and citybelonging levels
}
\description{
This function takes in a dataframe of individuals survey data containing gender, race, age info as in indicator matrix form.
takes n samples, perform raking algorism on each sample to get the weights, use the weights to get the percentage for each levels
in education and citybelongings.Finally get the mean, variance, CIs for the levels percentages using the samples.
}
