% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWeights.R
\name{getWeights}
\alias{getWeights}
\title{Return weights calculated using raking algorism}
\usage{
getWeights(inds, cons, feat, tols = 0.0115)
}
\arguments{
\item{inds}{a dataframe of individual survey data, in the form of indicator matrix, each row represent an individual}

\item{cons}{a dataframe of census constrain data, each row is a different census region, each column represent a census category.}

\item{feat}{a numerical vector representing the corresponding columns for each feature in the inds dataframe}

\item{tols}{tolerence for the stopping point of the iteration of the raking algorism}
}
\value{
A dataframe weights
}
\description{
This function takes in a dataframe of individual survey data, each row represents an individual.
a dataframe of census constrain data and perform raking algorism to get the weights for each
individual based on the census constrain
}
