% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rakeWeights.R
\name{rakeWeights}
\alias{rakeWeights}
\title{Get weights for a single feature}
\usage{
rakeWeights(ind, con, n_zone, n_ind, n_start, n_cons, wt, ind_agg, count,
  tol = 0.0115)
}
\value{
A list of the resulting weights, number of total iteration
each iteration will compare the difference between the current weights and the previous weights, if there is
practically no difference the number of iteration will be overwrote to 1000000001 to get out ot the iteration
}
\description{
This function takes in the indicator matrix for one feature and the census data.
using the raking algorism to calculate the weights for the feature according to the census constrain
}
