% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raking-package.R
\docType{package}
\name{raking-package}
\alias{raking-package}
\title{raking: Optimal Representative Sample Weighting}
\description{
Calibrates survey weights to known population targets using regularized raking, following Barratt, Angeris and Boyd (2021) "Optimal Representative Sample Weighting". Built on S7 classes with vectorized operations.
}
\author{
\strong{Maintainer}: wkt \email{wktc12@gmail.com}

}
\keyword{internal}
