% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{runExample}
\alias{runExample}
\title{Run shiny example}
\usage{
runExample()
}
\description{
This function will run the shiny app to interactively reprocess the survey input, 
get the census constrain data and get the weights
}
