% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSimulation.R
\name{runSimulation}
\alias{runSimulation}
\title{use getSim to run the simulation}
\usage{
runSimulation()
}
\description{
This function will run 2 simulations and print out the weighted mean/variance/CIs of the different levels of education and citybelongs 
simulation 1: take a random sample of 1000 individuals
simulation 2: take a biased sample of 1000 individuals, with 70% femals and 30% males
}
